feat(compiler): added initial ast generation (#15)
Okay this change to the compiler no longer spits out the parse tree when you run `cargo run` now we have the beginnings of our rust parser. This guy is responsible for reading the tree-sitter parse tree and then converting it to our AST for further compilation/interpretation. 

currently the AST only has a scope node but more will come as we add new grammar

Sample output
<img width="633" alt="image" src="https://github.com/titan-architecture/titan/assets/45375125/27804dbb-81e6-40ff-9a3c-72d48d675fb4">